AI/pyproject.toml
[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "intellibus-kids"
version = "0.1.0"
description = "AI healthcare assistant for children"
authors = [
    {name = "IntellibusDev"}
]
requires-python = ">=3.9"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "langchain>=0.1.0",
    "langchain-google-genai",
    "langgraph>=0.0.20",
    "pydantic>=2.0.0",
    "python-dotenv",
    "pinecone-client",
    "langfuse",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "ruff>=0.0.270",
]

[tool.pytest.ini_options]
testpaths = [".spec"]

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "B"]
ignore = []

[tool.setuptools]
packages = ["api", "config", "models", "services"] 